+++ SwiftUI_Pull_to_Refresh/AppHostingController.swift
@@ -1,0 +1,16 @@
+//
+//  AppHostingController.swift
+//  Sudoku
+//
+//  Created by Geri Borbás on 03/09/2021.
+//
+
+import SwiftUI
+
+
+class AppHostingController: UIHostingController<CitiesView> {
+	
+	override var prefersHomeIndicatorAutoHidden: Bool {
+		true
+	}
+}