+++ SwiftUI_Pull_to_Refresh/SceneDelegate.swift
@@ -21,9 +21,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use a UIHostingController as window root view controller.
         if let windowScene = scene as? UIWindowScene {
             let window = UIWindow(windowScene: windowScene)
-            window.rootViewController = UIHostingController(
-				rootView: WeatherView()
-					.preferredColorScheme(.dark)
+            window.rootViewController = AppHostingController(
+				rootView: CitiesView()
 			)
             self.window = window
             window.makeKeyAndVisible()